requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
cadquery>=2.5.0
numpy>=1.25.0
scipy>=1.11.0
shapely>=2.0.0
pydantic>=1.10,<3
pyyaml>=6.0

# DeepSeek LangChain 集成
langchain-deepseek>=0.1.0


